// src/shared/types/quran.ts
export interface Surah {
  nomor: number;
  nama: string;
  namaLatin: string;
  jumlahAyat: number;
  tempatTurun: string;
  arti: string;
  deskripsi: string;
  audioFull: {
    "01": string;
    "02": string;
    "03": string;
    "04": string;
    "05": string;
  };
}

export interface Ayat {
  nomorAyat: number;
  teksArab: string;
  teksLatin: string;
  teksIndonesia: string;
  audio: {
    "01": string;
    "02": string;
    "03": string;
    "04": string;
    "05": string;
  };
}

export interface SurahDetail extends Surah {
  ayat: Ayat[];
  suratSelanjutnya: false | Surah;
  suratSebelomnya: false | Surah;
}